feat(ui): content-type navigation with campaigns (replaces flows sidebar)
- Sidebar: CONTENT section with All, Articles, Social Posts, Newsletters, Campaigns
- Remove flows list from sidebar (doesn't scale)
- Content sub-pages: /content/articles, /content/social, /content/newsletters
- Campaigns page: /content/campaigns (replaces /flows)
- Campaign detail: /content/campaigns/[id] (replaces /flows/[id])
- Channel tabs only shown on "All" view (no duplication with sidebar nav)
- Dynamic titles per content type page
- Redirects: /flows → /content/campaigns, /flows/[id] → /content/campaigns/[id]
