+++ frontend/src/app/content/articles/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import ContentLibraryPage from "../page";
+
+export default function ArticlesPage() {
+  return <ContentLibraryPage contentTypeFilter="article" hideChannelTabs />;
+}
