+++ frontend/src/app/content/newsletters/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import ContentLibraryPage from "../page";
+
+export default function NewslettersPage() {
+  return <ContentLibraryPage contentTypeFilter="newsletter" hideChannelTabs />;
+}
