+++ frontend/src/app/content/social/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import ContentLibraryPage from "../page";
+
+export default function SocialPage() {
+  return <ContentLibraryPage contentTypeFilter="social" hideChannelTabs />;
+}
