+++ frontend/src/components/sidebar.tsx
@@ -12,6 +12,9 @@ import {
   Library,
   BarChart3,
   Brain,
+  FileText,
+  Layers,
+  Mail,
   MessageSquare,
   TrendingUp,
   Users,
@@ -176,7 +179,7 @@ export function Sidebar() {
   return (
     <aside className="flex h-screen w-64 flex-col border-r bg-sidebar text-sidebar-foreground">
       {/* Logo */}
-      <Link href="/flows" className="flex items-center gap-2.5 border-b px-4 py-4 hover:bg-muted/50 transition-colors">
+      <Link href="/content" className="flex items-center gap-2.5 border-b px-4 py-4 hover:bg-muted/50 transition-colors">
         <Image src="/logo.png" alt="FlowBoost" width={28} height={28} className="rounded-md" />
         <span className="text-lg font-semibold">flowboost</span>
       </Link>
@@ -222,18 +225,30 @@ export function Sidebar() {
           <CalendarDays className="h-4 w-4" />
           Calendar
         </Link>
-        <Link
-          href="/content"
-          className={cn(
-            "flex items-center gap-3 rounded-md px-3 py-2 text-sm font-medium transition-colors",
-            contentActive
-              ? "bg-sidebar-accent text-sidebar-accent-foreground"
-              : "text-sidebar-foreground/70 hover:bg-sidebar-accent/50 hover:text-sidebar-foreground"
-          )}
-        >
-          <Library className="h-4 w-4" />
-          Content
-        </Link>
+      </div>
+
+      {/* Content Section */}
+      <div className="px-3 pt-2 pb-1">
+        <div className="px-3 pb-2">
+          <span className="text-xs font-semibold uppercase tracking-wider text-muted-foreground">Content</span>
+        </div>
+        <div className="space-y-0.5">
+          <Link href="/content" className={cn("flex items-center gap-3 rounded-md px-3 py-2 text-sm font-medium transition-colors", pathname === "/content" ? "bg-sidebar-accent text-sidebar-accent-foreground" : "text-sidebar-foreground/70 hover:bg-sidebar-accent/50 hover:text-sidebar-foreground")}>
+            <Library className="h-4 w-4" />All
+          </Link>
+          <Link href="/content/articles" className={cn("flex items-center gap-3 rounded-md px-3 py-2 text-sm font-medium transition-colors", pathname === "/content/articles" ? "bg-sidebar-accent text-sidebar-accent-foreground" : "text-sidebar-foreground/70 hover:bg-sidebar-accent/50 hover:text-sidebar-foreground")}>
+            <FileText className="h-4 w-4" />Articles
+          </Link>
+          <Link href="/content/social" className={cn("flex items-center gap-3 rounded-md px-3 py-2 text-sm font-medium transition-colors", pathname === "/content/social" ? "bg-sidebar-accent text-sidebar-accent-foreground" : "text-sidebar-foreground/70 hover:bg-sidebar-accent/50 hover:text-sidebar-foreground")}>
+            <MessageSquare className="h-4 w-4" />Social Posts
+          </Link>
+          <Link href="/content/newsletters" className={cn("flex items-center gap-3 rounded-md px-3 py-2 text-sm font-medium transition-colors", pathname === "/content/newsletters" ? "bg-sidebar-accent text-sidebar-accent-foreground" : "text-sidebar-foreground/70 hover:bg-sidebar-accent/50 hover:text-sidebar-foreground")}>
+            <Mail className="h-4 w-4" />Newsletters
+          </Link>
+          <Link href="/content/campaigns" className={cn("flex items-center gap-3 rounded-md px-3 py-2 text-sm font-medium transition-colors", pathname.startsWith("/content/campaigns") ? "bg-sidebar-accent text-sidebar-accent-foreground" : "text-sidebar-foreground/70 hover:bg-sidebar-accent/50 hover:text-sidebar-foreground")}>
+            <Layers className="h-4 w-4" />Campaigns
+          </Link>
+        </div>
       </div>
 
       {/* Intelligence Section */}
@@ -281,127 +296,8 @@ export function Sidebar() {
         </div>
       </div>
 
-      {/* Flows Section */}
-      <div className="flex-1 flex flex-col min-h-0 px-3 pt-3">
-        <div className="flex items-center justify-between px-3 pb-2">
-          <span className="text-xs font-semibold uppercase tracking-wider text-muted-foreground">Flows</span>
-          <button
-            type="button"
-            onClick={handleOpenOnboarding}
-            className="text-muted-foreground hover:text-foreground transition-colors"
-            title="New Flow"
-          >
-            <Plus className="h-3.5 w-3.5" />
-          </button>
-        </div>
-
-        {/* Flow List */}
-        <div className="flex-1 overflow-y-auto space-y-0.5">
-          {activeFlows.map((flow) => {
-            const isActive = pathname === `/flows/${flow.id}`;
-            const isRenaming = renamingId === flow.id;
-
-            if (isRenaming) {
-              return (
-                <div key={flow.id} className="rounded-md bg-sidebar-accent px-3 py-1.5">
-                  <input
-                    value={renameValue}
-                    onChange={(e) => setRenameValue(e.target.value)}
-                    onBlur={() => handleRenameFlow(flow.id)}
-                    onKeyDown={(e) => { if (e.key === "Enter") handleRenameFlow(flow.id); if (e.key === "Escape") setRenamingId(null); }}
-                    autoFocus
-                    className="w-full text-sm bg-transparent outline-none caret-primary"
-                  />
-                </div>
-              );
-            }
-
-            return (
-              <div key={flow.id} className="group relative">
-                <Link
-                  href={`/flows/${flow.id}`}
-                  className={cn(
-                    "flex items-center gap-2 rounded-md px-3 py-1.5 text-sm transition-colors pr-8",
-                    isActive
-                      ? "bg-sidebar-accent text-sidebar-accent-foreground"
-                      : "text-sidebar-foreground/70 hover:bg-sidebar-accent/50 hover:text-sidebar-foreground"
-                  )}
-                >
-                  <span className="flex-1 truncate">{flow.title}</span>
-                  {flow.source === "pipeline" && <Sparkles className="h-3 w-3 shrink-0 text-violet-500" />}
-                </Link>
-                <DropdownMenu>
-                  <DropdownMenuTrigger asChild>
-                    <button className="absolute right-1 top-1/2 -translate-y-1/2 p-1 rounded-md hover:bg-sidebar-accent opacity-0 group-hover:opacity-100 data-[state=open]:opacity-100 transition-opacity text-muted-foreground">
-                      <MoreHorizontal className="h-3.5 w-3.5" />
-                    </button>
-                  </DropdownMenuTrigger>
-                  <DropdownMenuContent align="start" side="right">
-                    <DropdownMenuItem onClick={() => { setRenamingId(flow.id); setRenameValue(flow.title); }}>
-                      <Pencil className="mr-2 h-3.5 w-3.5" />Rename
-                    </DropdownMenuItem>
-                    <DropdownMenuItem onClick={() => handleArchiveFlow(flow.id)}>
-                      <Archive className="mr-2 h-3.5 w-3.5" />Archive
-                    </DropdownMenuItem>
-                    <DropdownMenuSeparator />
-                    <DropdownMenuItem className="text-destructive" onClick={() => setDeleteFlowId(flow.id)}>
-                      <Trash2 className="mr-2 h-3.5 w-3.5" />Delete
-                    </DropdownMenuItem>
-                  </DropdownMenuContent>
-                </DropdownMenu>
-              </div>
-            );
-          })}
-
-          {activeFlows.length === 0 && (
-            <p className="px-3 py-2 text-xs text-muted-foreground">No active flows</p>
-          )}
-
-          {/* Archive Toggle — always visible */}
-          <button
-            type="button"
-            onClick={() => setShowArchive(!showArchive)}
-            className="flex items-center gap-2 w-full rounded-md px-3 py-1.5 text-xs text-muted-foreground hover:text-foreground transition-colors mt-3"
-          >
-            <Archive className="h-3 w-3" />
-            Archive{archivedFlows.length > 0 && ` (${archivedFlows.length})`}
-            <ChevronDown className={cn("h-3 w-3 ml-auto transition-transform", showArchive && "rotate-180")} />
-          </button>
-
-          {showArchive && (
-            archivedFlows.length === 0 ? (
-              <p className="px-3 py-1.5 text-xs text-muted-foreground/50">No archived flows</p>
-            ) : (
-              archivedFlows.map((flow) => (
-                <div key={flow.id} className="group relative">
-                  <Link
-                    href={`/flows/${flow.id}`}
-                    className="flex items-center gap-2 rounded-md px-3 py-1.5 text-xs text-muted-foreground hover:text-foreground transition-colors pr-8"
-                  >
-                    <span className="truncate">{flow.title}</span>
-                  </Link>
-                  <DropdownMenu>
-                    <DropdownMenuTrigger asChild>
-                      <button className="absolute right-1 top-1/2 -translate-y-1/2 p-1 rounded-md hover:bg-sidebar-accent opacity-0 group-hover:opacity-100 transition-opacity text-muted-foreground">
-                        <MoreHorizontal className="h-3.5 w-3.5" />
-                      </button>
-                    </DropdownMenuTrigger>
-                    <DropdownMenuContent align="start" side="right">
-                      <DropdownMenuItem onClick={() => handleUnarchiveFlow(flow.id)}>
-                        <Archive className="mr-2 h-3.5 w-3.5" />Unarchive
-                      </DropdownMenuItem>
-                      <DropdownMenuSeparator />
-                      <DropdownMenuItem className="text-destructive" onClick={() => setDeleteFlowId(flow.id)}>
-                        <Trash2 className="mr-2 h-3.5 w-3.5" />Delete
-                      </DropdownMenuItem>
-                    </DropdownMenuContent>
-                  </DropdownMenu>
-                </div>
-              ))
-            )
-          )}
-        </div>
-      </div>
+      {/* Spacer */}
+      <div className="flex-1" />
 
       {/* Media + Monitor */}
       <div className="px-3 pb-1 space-y-0.5">
